Add @generated tag to generated bindata file
Currently Phabricator will ignore files with the string `@generated`
somewhere in the file. In order to not conflict with githubs definition
of a generated file I appended it to the end of the line. Looks a little
funky but shouldn't affect anything. I can also add this in a separate
line if we like that more.