Make *Bytes() function public
Our use case is we need to send data files and scripts to run on remote
servers.  In this case the specific file used depends on the step and is
baked into the code.  If we were to use Asset("...") to reference the
file there's a chance that a typo could cause errors at runtime.

By exposing the *Bytes() function we can reference the data and the
compiler will alert us if there is a mistake.

There are some caveats with the approach in this PR:
1. If the file begins with a number the resulting files will be private.
2. If there are files with conflicting names they may
   nondeterministically get numeric suffixes.

I think these are probably acceptable.  If the scripts, data files, etc.
were provided by a third party we'd probably be accessing them
programmatically.  Since we're providing the scripts ourselves we can
explicitly give them non-conflicting names.